# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScorpStudy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List quiz results
      responses:
        "200":
          description: List of quiz results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizResult"
    post:
      operationId: saveQuizResult
      tags: [quizzes]
      summary: Save quiz result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Saved quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
  /quizzes/{id}:
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete a quiz result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Flashcards ──────────────────────────────────────────────────────────────
  /flashcards:
    get:
      operationId: listFlashcardSets
      tags: [flashcards]
      summary: List flashcard sets
      responses:
        "200":
          description: List of flashcard sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlashcardSet"
    post:
      operationId: saveFlashcardSet
      tags: [flashcards]
      summary: Save flashcard set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardInput"
      responses:
        "201":
          description: Saved flashcard set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardSet"
  /flashcards/{id}:
    delete:
      operationId: deleteFlashcardSet
      tags: [flashcards]
      summary: Delete a flashcard set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Summaries ───────────────────────────────────────────────────────────────
...
  /images:
    get:
      operationId: listImages
      tags: [images]
      summary: List generated images
      responses:
        "200":
          description: List of generated images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneratedImage"
    post:
      operationId: saveImage
      tags: [images]
      summary: Save a generated image record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageInput"
      responses:
        "201":
          description: Saved image record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"
  /images/{id}:
    delete:
      operationId: deleteImage
      tags: [images]
      summary: Delete an image record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── AI Endpoints ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    QuizQuestion:
      type: object
      required: [question, options, correctAnswer, explanation]
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
    QuizResult:
      type: object
      required: [id, topic, score, total, questions, createdAt]
      properties:
        id:
          type: integer
        topic:
          type: string
        score:
          type: integer
        total:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        createdAt:
          type: string
    QuizInput:
      type: object
      required: [topic, score, total, questions]
      properties:
        topic:
          type: string
        score:
          type: integer
        total:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"

    # ── Flashcards ──────────────────────────────────────────────────────────
    Flashcard:
      type: object
      required: [front, back]
      properties:
        front:
          type: string
        back:
          type: string
    FlashcardSet:
      type: object
      required: [id, topic, cards, createdAt]
      properties:
        id:
          type: integer
        topic:
          type: string
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
        createdAt:
          type: string
    FlashcardInput:
      type: object
      required: [topic, cards]
      properties:
        topic:
          type: string
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"

    # ── Summaries ──────────────────────────────────────────────────────────
...
    GeneratedImage:
      type: object
      required: [id, prompt, imageUrl, style, createdAt]
      properties:
        id:
          type: integer
        prompt:
          type: string
        imageUrl:
          type: string
        style:
          type: string
        createdAt:
          type: string
    ImageInput:
      type: object
      required: [prompt, imageUrl, style]
      properties:
        prompt:
          type: string
        imageUrl:
          type: string
        style:
          type: string

    # ── AI Requests/Responses ───────────────────────────────────────────────
...