# .migration-backup/lib/api-zod/src/generated/types/flashcardInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ScorpStudy API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Flashcard } from './flashcard';

export interface FlashcardInput {
  topic: string;
  cards: Flashcard[];
}