# .migration-backup/lib/api-zod/src/generated/types/generatedImage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ScorpStudy API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedImage {
  id: number;
  prompt: string;
  imageUrl: string;
  style: string;
  createdAt: string;
}